Moved key generator to its own file
