+++ src/ComponentMap.hpp
@@ -10,6 +10,7 @@
 #include "YAGL_Algorithms.hpp"
 #include "YAGL_Graph.hpp"
 #include "AnalyzedGrammar.hpp"
+#include "KeyGenerator.hpp"
 
 namespace DGGML
 {
@@ -30,22 +31,6 @@ namespace DGGML
             os << "Intermediate";
         return os;
     }
-
-    //We need matching rewrite functions for every enum class member
-
-    template<typename KeyType>
-    struct KeyGenerator 
-    {
-        using key_type = KeyType;
-        
-        key_type current_key;
-
-        KeyGenerator(KeyType key) : current_key(key) {}
-
-        KeyGenerator() : current_key(0) {}
-
-        key_type get_key() { return current_key++; }
-    };
 
     template <typename KeyType>
     struct RuleInstType {
